+++ areas/appconfig/src/AzureMcp.AppConfig/Options/AppConfigOptionDefinitions.cs
@@ -9,6 +9,8 @@ public static class AppConfigOptionDefinitions
     public const string KeyName = "key";
     public const string ValueName = "value";
     public const string LabelName = "label";
+    public const string ContentTypeName = "content-type";
+    public const string TagsName = "tags";
 
     public static readonly Option<string> Account = new(
         $"--{AccountName}",
@@ -40,6 +42,23 @@ public static class AppConfigOptionDefinitions
     )
     {
         IsRequired = false
+    };
+
+    public static readonly Option<string> ContentType = new(
+        $"--{ContentTypeName}",
+        "The content type of the configuration value. This is used to indicate how the value should be interpreted or parsed."
+    )
+    {
+        IsRequired = false
+    };
+
+    public static readonly Option<string[]> Tags = new(
+        $"--{TagsName}",
+        "The tags to associate with the configuration key. Tags should be in the format 'key=value'. Multiple tags can be specified."
+    )
+    {
+        IsRequired = false,
+        AllowMultipleArgumentsPerToken = true
     };
 
     public static class KeyValueList