+++ areas/appconfig/src/AzureMcp.AppConfig/Options/KeyValue/KeyValueSetOptions.cs
@@ -9,4 +9,7 @@ public class KeyValueSetOptions : BaseKeyValueOptions
 {
     [JsonPropertyName(AppConfigOptionDefinitions.ValueName)]
     public string? Value { get; set; }
+
+    [JsonPropertyName(AppConfigOptionDefinitions.TagsName)]
+    public string[]? Tags { get; set; }
 }